# OpenAPI/BastionManager.json
{
  "openapi": "3.1.0",
  "x-stoplight": {
    "id": "5rtjmrxf7j4ik"
  },
  "info": {
    "title": "BastionManager",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/bastions": {
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bastion"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-bastion",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "affichage différent selon user ou admin"
          }
        ],
        "description": "affiche la liste complète/d'appartenance selon si l'on est admin/user"
      },
      "post": {
        "summary": "",
        "operationId": "post-bastion",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConfigAgent"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BastionModification"
              }
            }
          },
          "description": ""
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "admin seulement"
          }
        ],
        "description": "création d'un bastion"
      },
      "parameters": []
    },
    "/bastions/{bastion_id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "bastion_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bastion"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-bastion-bastion_id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "affiche un bastion correspondant à l'URL"
          }
        ],
        "description": "affiche le bastion d'id {bastion_id}"
      },
      "patch": {
        "summary": "",
        "operationId": "patch-bastion-bastion_id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bastion"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BastionModification"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "admin seulement"
          }
        ],
        "description": "modifie le bastion"
      },
      "delete": {
        "summary": "",
        "operationId": "delete-bastion-bastion_id",
        "responses": {
          "200": {
            "description": "\"supprimé\""
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "admin seulement"
          }
        ],
        "description": "supprime le bastion"
      }
    },
    "/bastions/{bastion_id}/users": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "bastion_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-bastion-bastion_id-users",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "admin seulement"
          }
        ],
        "description": "affiche la liste d'utilisateur sur le bastion"
      },
      "post": {
        "summary": "",
        "operationId": "post-bastion-bastion_id-users",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConfigUser"
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModification"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "admin seulement"
          }
        ],
        "description": "ajout d'un utilisateur au bastion"
      }
    },
    "/bastions/{bastion_id}/users/{user_id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "bastion_id",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "user_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-bastion-bastion_id-users-user_id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "admin seulement"
          }
        ],
        "description": "affiche la description d'un utilisateur"
      },
      "delete": {
        "summary": "",
        "operationId": "delete-bastion-bastion_id-users-user_id",
        "responses": {
          "200": {
            "description": "\"supprimé\""
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "admin seulement"
          }
        ],
        "description": "supprime l'accés de l'utilisateur"
      }
    },
    "/bastions/{bastion_id}/users/{user_id}/generate_wireguard": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "bastion_id",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "user_id",
          "in": "path",
          "required": true
        }
      ],
      "post": {
        "summary": "",
        "operationId": "post-bastions-bastion_id-users-user_id-generate_wireguard",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConfigClient"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "cookie",
            "name": "id",
            "description": "pour utilisateur uniquement"
          }
        ]
      }
    },
    "/bastions/{bastion_id}/ressources": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "bastion_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ressource"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-bastions-bastion_id-ressources",
        "description": "affiches l'ensemble des ressources accessible par le bastion"
      },
      "post": {
        "summary": "",
        "operationId": "post-bastions-bastion_id-ressources",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ressource"
                }
              }
            }
          }
        },
        "description": "crée une nouvelle ressource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "application/xml": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            },
            "text/html": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "description": ""
        }
      }
    },
    "/bastions/{bastion_id}/ressources/{ressource_id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "bastion_id",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "ressource_id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "supprimé"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-bastions-bastion_id-ressources-ressource_id",
        "description": "affiche une ressource"
      },
      "delete": {
        "summary": "",
        "operationId": "delete-bastions-bastion_id-ressources-ressource_id",
        "responses": {
          "200": {
            "description": "supprimé"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "description": "enlève la ressource en bdd"
      }
    }
  },
  "components": {
    "schemas": {
      "Bastion": {
        "title": "Bastion",
        "x-stoplight": {
          "id": "tddw6lrvd6ztr"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "subnet_cidr": {
            "type": "string"
          },
          "agent_endpoint": {
            "type": "string"
          },
          "pubkey": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "net_id": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "subnet_cidr",
          "agent_endpoint",
          "pubkey",
          "port",
          "net_id"
        ]
      },
      "Users": {
        "title": "User",
        "x-stoplight": {
          "id": "e6423xqu0qwe5"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "bastion_id": {
            "type": "integer"
          },
          "wireguard": {
            "type": "boolean"
          },
          "net_id": {
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "bastion_id",
          "wireguard",
          "net_id"
        ]
      },
      "ConfigAgent": {
        "title": "ConfigAgent",
        "x-stoplight": {
          "id": "0ms8xmek5bsnl"
        },
        "type": "object",
        "properties": {
          "privkey": {
            "type": "string",
            "description": "de l'agent"
          },
          "pubkey": {
            "type": "string",
            "description": "du bastion"
          }
        },
        "required": [
          "privkey",
          "pubkey"
        ]
      },
      "BastionModification": {
        "title": "BastionModification",
        "x-stoplight": {
          "id": "v5012yxuuz7ey"
        },
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "subnet_cidr": {
            "type": "string"
          },
          "agent_endpoint": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "subnet_cidr",
          "agent_endpoint"
        ]
      },
      "ConfigUser": {
        "title": "ConfigUser",
        "x-stoplight": {
          "id": "pafnkoi4az70m"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "net_id": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "net_id"
        ]
      },
      "UserModification": {
        "title": "UserModification",
        "x-stoplight": {
          "id": "e5ckopu2vd3q9"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
      "ConfigClient": {
        "title": "ConfigClient",
        "x-stoplight": {
          "id": "xaymm88bi2der"
        },
        "type": "object",
        "properties": {
          "client_private_key": {
            "type": "string"
          },
          "client_address": {
            "type": "string"
          },
          "bastion_public_key": {
            "type": "string"
          },
          "bastion_endpoint": {
            "type": "string"
          },
          "subnet_cidr": {
            "type": "string"
          }
        }
      },
      "Ressource": {
        "title": "Ressource",
        "x-stoplight": {
          "id": "d0bmhpfa1r0ur"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "id_bastion": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "id_wireguard": {
            "type": "string"
          },
          "id_ssh": {
            "type": "string"
          },
          "id_k8s": {
            "type": "string"
          }
        }
      },
      "Session": {
        "title": "Session",
        "x-stoplight": {
          "id": "lzh3o37p0313i"
        },
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "id_ressource": {
            "type": "string"
          },
          "id_user": {
            "type": "string"
          },
          "temps_fin": {
            "type": "integer"
          }
        }
      },
      "WireguardRessource": {
        "title": "WireguardSession",
        "x-stoplight": {
          "id": "zje9fgakvkvo6"
        },
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "target_cidr": {
            "type": "string"
          }
        }
      },
      "SSHRessource": {
        "title": "SSHSession",
        "x-stoplight": {
          "id": "iv8pont40w0et"
        },
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "ip_machine": {
            "type": "string"
          }
        }
      },
      "K8SRessource": {
        "title": "K8SRessource",
        "x-stoplight": {
          "id": "2lxax15vi5fyu"
        },
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "ip_cluster": {
            "type": "string"
          }
        }
      }
    }
  }
}